Fix plugin crashing bug for non-primitive proto types, refactoring
